+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- All nodes with `indentationLevel` attribute can be indented, instead of only hardcoded nodes.
 
 ### Fixed
 - All Paragraphs are now part of the group `paragraphGroup`
@@ -23,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumps `prettier` from 3.0.0 to 3.0.1
 - Bumps `prosemirror-transform` from 1.7.3 to 1.7.4
 - Bumps `@codemirror/view` from 6.15.3 to 6.16.0
+
+### Breaking
+- remove option `allowedTypes`. Instead all nodes with attribute `indentationLevel` can be indented.
 
 ## [4.2.0] - 2023-07-29
 ### Dependencies